tog patch

+++ demo_patch.py
@@ -277,7 +277,7 @@ print(x_adv)
 print(x_adv.shape)
 print(x_adv.dtype)
 # dtype =  torch.cuda.FloatTensor if torch.cuda.is_available() else torch.FloatTensor
-dtype = torch.FloatTensor
+dtype = torch.DoubleTensor
 detections_adv = detector.detect(x_adv, conf_threshold=detector.confidence_thresh_default)
 detections_rand = detector.detect(x_rand, conf_threshold=detector.confidence_thresh_default)
 x_ad = torch.tensor(x_adv)
